Remove test framework stuff into common module.
