+++ substrate/runtime/executive/Cargo.toml
@@ -14,7 +14,6 @@ substrate-runtime-primitives = { path = "../primitives", default_features = fals
 substrate-runtime-system = { path = "../system", default_features = false }
 
 [dev-dependencies]
-serde_derive = { version = "1.0" }
 substrate-primitives = { path = "../../primitives" }
 substrate-runtime-session = { path = "../session" }
 substrate-runtime-staking = { path = "../staking" }
