+++ substrate/runtime/executive/src/lib.rs
@@ -24,7 +24,6 @@ extern crate substrate_runtime_io as runtime_io;
 extern crate substrate_codec as codec;
 extern crate substrate_runtime_primitives as primitives;
 extern crate substrate_runtime_system as system;
-#[cfg(test)] #[macro_use] extern crate serde_derive;
 #[cfg(test)] extern crate substrate_primitives;
 #[cfg(test)] extern crate substrate_runtime_consensus as consensus;
 #[cfg(test)] extern crate substrate_runtime_session as session;
@@ -35,7 +34,7 @@ extern crate substrate_runtime_system as system;
 use rstd::prelude::*;
 use rstd::marker::PhantomData;
 use runtime_io::Hashing;
-use primitives::{Zero, One, Headery, Blocky, Checkable, Applyable, CheckEqual, Executable};
+use primitives::{Zero, One, Headery, Blocky, Checkable, Applyable, CheckEqual, Executable, MakePayment};
 use codec::Slicable;
 
 pub struct Executive<
@@ -43,8 +42,9 @@ pub struct Executive<
 	Checked,
 	System,
 	Block,
+	Payment,
 	Finalisation,
->(PhantomData<(Unchecked, Checked, System, Block, Finalisation)>);
+>(PhantomData<(Unchecked, Checked, System, Block, Payment, Finalisation)>);
 
 impl<
 	Unchecked: Checkable<
@@ -59,8 +59,9 @@ impl<
 		Extrinsic = Unchecked,
 		Header = System::Header
 	>,
+	Payment: MakePayment<System::AccountId>,
 	Finalisation: Executable,
-> Executive<Unchecked, Checked, System, Block, Finalisation> {
+> Executive<Unchecked, Checked, System, Block, Payment, Finalisation> {
 	/// Start the execution of a particular block.
 	pub fn initialise_block(header: &System::Header) {
 		<system::Module<System>>::initialise(header.number(), header.parent_hash(), header.extrinsics_root());
@@ -134,6 +135,9 @@ impl<
 			<system::Module<System>>::inc_account_index(tx.sender());
 		}
 
+		// pay any fees.
+		Payment::make_payment(tx.sender());
+
 		// decode parameters and dispatch
 		tx.apply();
 	}
@@ -162,141 +166,12 @@ impl<
 mod tests {
 	use super::*;
 	use staking::Call;
-	use codec::{Slicable, Input};
 	use runtime_io::with_externalities;
-	use runtime_support::AuxDispatchable;
 	use substrate_primitives::H256;
-	use primitives::{Checkable, Applyable, HasPublicAux, Identity, Headery, Blocky, Digesty, MakeTestExternalities};
-
-	#[derive(Default, PartialEq, Eq, Clone, Serialize, Debug)]
-	pub struct Digest {
-		pub logs: Vec<u64>,
-	}
-	impl Slicable for Digest {
-		fn decode<I: Input>(input: &mut I) -> Option<Self> {
-			Vec::<u64>::decode(input).map(|logs| Digest { logs })
-		}
-		fn using_encoded<R, F: FnOnce(&[u8]) -> R>(&self, f: F) -> R {
-			self.logs.using_encoded(f)
-		}
-	}
-	impl Digesty for Digest {
-		type Item = u64;
-		fn push(&mut self, item: Self::Item) {
-			self.logs.push(item);
-		}
-	}
-
-	#[derive(PartialEq, Eq, Clone, Serialize, Debug)]
-	#[serde(rename_all = "camelCase")]
-	#[serde(deny_unknown_fields)]
-	pub struct Header {
-		pub parent_hash: H256,
-		pub number: u64,
-		pub state_root: H256,
-		pub extrinsics_root: H256,
-		pub digest: Digest,
-	}
-	impl Slicable for Header {
-		fn decode<I: Input>(input: &mut I) -> Option<Self> {
-			Some(Header {
-				parent_hash: Slicable::decode(input)?,
-				number: Slicable::decode(input)?,
-				state_root: Slicable::decode(input)?,
-				extrinsics_root: Slicable::decode(input)?,
-				digest: Slicable::decode(input)?,
-			})
-		}
-
-		fn encode(&self) -> Vec<u8> {
-			let mut v = Vec::new();
-			self.parent_hash.using_encoded(|s| v.extend(s));
-			self.number.using_encoded(|s| v.extend(s));
-			self.state_root.using_encoded(|s| v.extend(s));
-			self.extrinsics_root.using_encoded(|s| v.extend(s));
-			self.digest.using_encoded(|s| v.extend(s));
-			v
-		}
-	}
-	impl Headery for Header {
-		type Number = u64;
-		type Hash = H256;
-		type Digest = Digest;
-		fn number(&self) -> &Self::Number { &self.number }
-		fn extrinsics_root(&self) -> &Self::Hash { &self.extrinsics_root }
-		fn state_root(&self) -> &Self::Hash { &self.state_root }
-		fn parent_hash(&self) -> &Self::Hash { &self.parent_hash }
-		fn digest(&self) -> &Self::Digest { &self.digest }
-		fn new(
-			number: Self::Number,
-			extrinsics_root: Self::Hash,
-			state_root: Self::Hash,
-			parent_hash: Self::Hash,
-			digest: Self::Digest
-		) -> Self {
-			Header {
-				number, extrinsics_root: extrinsics_root, state_root, parent_hash, digest
-			}
-		}
-	}
-
-	#[derive(PartialEq, Eq, Clone, Serialize, Debug)]
-	pub struct Block {
-		pub header: Header,
-		pub extrinsics: Vec<TestXt>,
-	}
-	impl Slicable for Block {
-		fn decode<I: Input>(input: &mut I) -> Option<Self> {
-			Some(Block {
-				header: Slicable::decode(input)?,
-				extrinsics: Slicable::decode(input)?,
-			})
-		}
-		fn encode(&self) -> Vec<u8> {
-			let mut v: Vec<u8> = Vec::new();
-			v.extend(self.header.encode());
-			v.extend(self.extrinsics.encode());
-			v
-		}
-	}
-	impl Blocky for Block {
-		type Extrinsic = TestXt;
-		type Header = Header;
-		fn header(&self) -> &Self::Header {
-			&self.header
-		}
-		fn extrinsics(&self) -> &[Self::Extrinsic] {
-			&self.extrinsics[..]
-		}
-		fn deconstruct(self) -> (Self::Header, Vec<Self::Extrinsic>) {
-			(self.header, self.extrinsics)
-		}
-	}
+	use primitives::{HasPublicAux, Identity, Headery, MakeTestExternalities};
+	use primitives::testing::{Digest, Header, Block};
 
 	pub struct Test;
-
-	#[derive(PartialEq, Eq, Clone, Serialize, Debug)]
-	pub struct TestXt((u64, u64, Call<Test>));
-	impl Slicable for TestXt {
-		fn decode<I: Input>(input: &mut I) -> Option<Self> {
-			Some(TestXt(Slicable::decode(input)?))
-		}
-		fn encode(&self) -> Vec<u8> {
-			self.0.encode()
-		}
-	}
-	impl Checkable for TestXt {
-		type Checked = Self;
-		fn check(self) -> Result<Self, Self> { Ok(self) }
-	}
-	impl Applyable for TestXt {
-		type AccountId = u64;
-		type Index = u64;
-		fn sender(&self) -> &u64 { &(self.0).0 }
-		fn index(&self) -> &u64 { &(self.0).1 }
-		fn apply(self) { <staking::Module<Test>>::make_payment(&(self.0).0); (self.0).2.dispatch(&(self.0).0) }
-	}
-
 	impl HasPublicAux for Test {
 		type PublicAux = u64;
 	}
@@ -316,17 +191,13 @@ mod tests {
 		type PublicAux = <Self as HasPublicAux>::PublicAux;
 		type ConvertAccountIdToSessionKey = Identity;
 	}
-	pub struct DummyContractAddressFor;
-	impl staking::ContractAddressFor<u64> for DummyContractAddressFor {
-		fn contract_address_for(_code: &[u8], origin: &u64) -> u64 {
-			origin + 1
-		}
-	}
 	impl staking::Trait for Test {
 		type Balance = u64;
-		type DetermineContractAddress = DummyContractAddressFor;
+		type DetermineContractAddress = staking::DummyContractAddressFor;
 	}
-	type Executive = super::Executive<TestXt, TestXt, Test, Block, ()>;
+
+	type TestXt = primitives::testing::TestXt<Call<Test>>;
+	type Executive = super::Executive<TestXt, TestXt, Test, Block<TestXt>, staking::Module<Test>, ()>;
 
 	#[test]
 	fn staking_balance_transfer_dispatch_works() {
@@ -340,7 +211,7 @@ mod tests {
 			bonding_duration: 0,
 			transaction_fee: 10,
 		}.test_externalities());
-		let xt = TestXt((1, 0, Call::transfer(2, 69)));
+		let xt = primitives::testing::TestXt((1, 0, Call::transfer(2, 69)));
 		with_externalities(&mut t, || {
 			Executive::initialise_block(&Header::new(1, H256::default(), H256::default(), [69u8; 32].into(), Digest::default()));
 			Executive::apply_extrinsic(xt);
