+++ substrate/runtime/primitives/Cargo.toml
@@ -6,17 +6,23 @@ authors = ["Parity Technologies <admin@parity.io>"]
 [dependencies]
 num-traits = { version = "0.2", default_features = false }
 integer-sqrt = "0.1.0"
-substrate-runtime-std = { path = "../../runtime-std", default_features = false }
-substrate-runtime-io = { path = "../../runtime-io", default_features = false }
+serde = { version = "1.0", optional = true }
+serde_derive = { version = "1.0", optional = true }
 substrate-codec = { path = "../../codec", default_features = false }
 substrate-primitives = { path = "../../primitives", default_features = false }
+substrate-runtime-std = { path = "../../runtime-std", default_features = false }
+substrate-runtime-io = { path = "../../runtime-io", default_features = false }
+substrate-runtime-support = { path = "../../runtime-support", default_features = false }
 
 [features]
 default = ["std"]
 std = [
 	"num-traits/std",
+	"serde",
+	"serde_derive",
 	"substrate-runtime-std/std",
 	"substrate-runtime-io/std",
+	"substrate-runtime-support/std",
 	"substrate-codec/std",
 	"substrate-primitives/std",
 ]
